planner: move PhysicalIndexJoin related logic into physical op dir
