+++ pkg/planner/core/plan_cost_ver1.go
@@ -452,92 +452,6 @@ func (p *PhysicalIndexScan) GetPlanCostVer1(_ property.TaskType, option *optimiz
 	return p.PlanCost, nil
 }
 
-// GetCost computes the cost of index join operator and its children.
-func (p *PhysicalIndexJoin) GetCost(outerCnt, innerCnt, outerCost, innerCost float64, costFlag uint64) float64 {
-	var cpuCost float64
-	sessVars := p.SCtx().GetSessionVars()
-	// Add the cost of evaluating outer filter, since inner filter of index join
-	// is always empty, we can simply tell whether outer filter is empty using the
-	// summed length of left/right conditions.
-	if len(p.LeftConditions)+len(p.RightConditions) > 0 {
-		cpuCost += sessVars.GetCPUFactor() * outerCnt
-		outerCnt *= cost.SelectionFactor
-	}
-	// Cost of extracting lookup keys.
-	innerCPUCost := sessVars.GetCPUFactor() * outerCnt
-	// Cost of sorting and removing duplicate lookup keys:
-	// (outerCnt / batchSize) * (batchSize * Log2(batchSize) + batchSize) * CPUFactor
-	batchSize := math.Min(float64(p.SCtx().GetSessionVars().IndexJoinBatchSize), outerCnt)
-	if batchSize > 2 {
-		innerCPUCost += outerCnt * (math.Log2(batchSize) + 1) * sessVars.GetCPUFactor()
-	}
-	// Add cost of building inner executors. CPU cost of building copTasks:
-	// (outerCnt / batchSize) * (batchSize * DistinctFactor) * CPUFactor
-	// Since we don't know the number of copTasks built, ignore these network cost now.
-	innerCPUCost += outerCnt * cost.DistinctFactor * sessVars.GetCPUFactor()
-	// CPU cost of building hash table for inner results:
-	// (outerCnt / batchSize) * (batchSize * DistinctFactor) * innerCnt * CPUFactor
-	innerCPUCost += outerCnt * cost.DistinctFactor * innerCnt * sessVars.GetCPUFactor()
-	innerConcurrency := float64(p.SCtx().GetSessionVars().IndexLookupJoinConcurrency())
-	cpuCost += innerCPUCost / innerConcurrency
-	// Cost of probing hash table in main thread.
-	numPairs := outerCnt * innerCnt
-	if p.JoinType == logicalop.SemiJoin || p.JoinType == logicalop.AntiSemiJoin ||
-		p.JoinType == logicalop.LeftOuterSemiJoin || p.JoinType == logicalop.AntiLeftOuterSemiJoin {
-		if len(p.OtherConditions) > 0 {
-			numPairs *= 0.5
-		} else {
-			numPairs = 0
-		}
-	}
-	if hasCostFlag(costFlag, costusage.CostFlagUseTrueCardinality) {
-		numPairs = getOperatorActRows(p)
-	}
-	probeCost := numPairs * sessVars.GetCPUFactor()
-	// Cost of additional concurrent goroutines.
-	cpuCost += probeCost + (innerConcurrency+1.0)*sessVars.GetConcurrencyFactor()
-	// Memory cost of hash tables for inner rows. The computed result is the upper bound,
-	// since the executor is pipelined and not all workers are always in full load.
-	memoryCost := innerConcurrency * (batchSize * cost.DistinctFactor) * innerCnt * sessVars.GetMemoryFactor()
-	// Cost of inner child plan, i.e, mainly I/O and network cost.
-	innerPlanCost := outerCnt * innerCost
-	if p.SCtx().GetSessionVars().CostModelVersion == 2 {
-		// IndexJoin executes a batch of rows at a time, so the actual cost of this part should be
-		//  `innerCostPerBatch * numberOfBatches` instead of `innerCostPerRow * numberOfOuterRow`.
-		// Use an empirical value batchRatio to handle this now.
-		// TODO: remove this empirical value.
-		batchRatio := 30.0
-		innerPlanCost /= batchRatio
-	}
-	return outerCost + innerPlanCost + cpuCost + memoryCost
-}
-
-// GetPlanCostVer1 calculates the cost of the plan if it has not been calculated yet and returns the cost.
-func (p *PhysicalIndexJoin) GetPlanCostVer1(taskType property.TaskType, option *optimizetrace.PlanCostOption) (float64, error) {
-	costFlag := option.CostFlag
-	if p.PlanCostInit && !hasCostFlag(costFlag, costusage.CostFlagRecalculate) {
-		return p.PlanCost, nil
-	}
-	outerChild, innerChild := p.Children()[1-p.InnerChildIdx], p.Children()[p.InnerChildIdx]
-	outerCost, err := outerChild.GetPlanCostVer1(taskType, option)
-	if err != nil {
-		return 0, err
-	}
-	innerCost, err := innerChild.GetPlanCostVer1(taskType, option)
-	if err != nil {
-		return 0, err
-	}
-	outerCnt := getCardinality(outerChild, costFlag)
-	innerCnt := getCardinality(innerChild, costFlag)
-	if hasCostFlag(costFlag, costusage.CostFlagUseTrueCardinality) && outerCnt > 0 {
-		innerCnt /= outerCnt // corresponding to one outer row when calculating IndexJoin costs
-		innerCost /= outerCnt
-	}
-	p.PlanCost = p.GetCost(outerCnt, innerCnt, outerCost, innerCost, costFlag)
-	p.PlanCostInit = true
-	return p.PlanCost, nil
-}
-
 // GetCost computes the cost of index merge join operator and its children.
 func (p *PhysicalIndexHashJoin) GetCost(outerCnt, innerCnt, outerCost, innerCost float64, costFlag uint64) float64 {
 	var cpuCost float64
